+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 Full changelog for the Transmute Coffee website.
 
-## v1.26.0 - New release - 2018-05-29
+## v1.27.0 - New release - 2018-05-29
 
 * New release, Quill PHP renderer, v3.11.0.
 * Demo page updated to use v3.11.0.
 * Navbar link for Quill PHP renderer now goes to detail page.
+
+## v1.26.0 - New release - 2018-05-21
+
+* New release, Quill PHP renderer, v3.10.1.
+* Updated PHP Quill Renderer detail page.
 
 ## v1.25.0 - Quill detail page - 2018-05-19